+++ src/MyMediaLite/Correlation/Jaccard.cs
@@ -54,7 +54,7 @@ namespace MyMediaLite.Correlation
 		}
 
 		///
-		protected override float ComputeCorrelationFromOverlap(float overlap, int count_x, int count_y)
+		protected override float ComputeCorrelationFromOverlap(float overlap, float count_x, float count_y)
 		{
 			if (overlap != 0)
 				return overlap / (count_x + count_y - overlap);
